+++ routers/crud_router.go
@@ -31,7 +31,6 @@ func init() {
 	beego.Router("/bee/actione/:app/:table/:view/:form/:id/:action", &controllers.CrudActionElementController{})
 	beego.Router("/bee/actionx/:app/:table/:view/:id/:action", &controllers.CrudActionAjaxController{})
 	beego.Router("/bee/ajax/:app/:table/:view/:form/:action", &controllers.CrudAjaxSqlController{}, "post:Post")
-	beego.Router("/bee/share/:appid/:shareid", &controllers.CrudShareController{}, "get:CrudShareApp")
 	beego.Router("/bee/search/:app/:table/:view", &controllers.CrudSearchController{}, "post:Post")
 	beego.Router("/bee/filter/:app/:table/:view", &controllers.CrudFilterController{}, "post:Post")
 	beego.Router("/bee/sort/:app/:table/:view", &controllers.CrudSortController{}, "post:Post")