+++ CHANGELOG.md
@@ -10,10 +10,13 @@ Historique des modifications
 - `deleted`
 - `changed`
 
+5.6.1 - 30 juillet 2025
+- `removed` suppression fonction de partage d'application
+- `removed` option Application.Target
+
 5.6.0 - 28 juillet 2025
 - `added` icone-name dans Action
 - `added` doc structure de dictionnaire
-
 
 5.5.5 - 25 juin 2025
 - `added` icone-about sur l'accueil et apropos